requirements.txt
slackclient
certifi
